+++ ethstorage/archiver/api.go
@@ -8,13 +8,10 @@ import (
 	"fmt"
 	"math/big"
 	"net/http"
-	"slices"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/crypto/kzg4844"
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethstorage/go-ethstorage/ethstorage"
@@ -164,78 +161,4 @@ func (a *API) buildSidecar(kvIndex uint64, kzgCommitment []byte, blobHash common
 		KZGCommitment: [48]byte(kzgCommitment),
 		KZGProof:      kzgProof,
 	}, nil
-}
-
-func validateBlockID(id string) *httpError {
-	if isHash(id) || isSlot(id) || isKnownIdentifier(id) {
-		return nil
-	}
-	return newBlockIdError(id)
-}
-
-func isHash(s string) bool {
-	if len(s) != 66 || !strings.HasPrefix(s, "0x") {
-		return false
-	}
-	_, err := hexutil.Decode(s)
-	return err == nil
-}
-
-func isSlot(id string) bool {
-	_, err := strconv.ParseUint(id, 10, 64)
-	return err == nil
-}
-
-func isKnownIdentifier(id string) bool {
-	return slices.Contains([]string{"genesis", "finalized", "head"}, id)
-}
-
-type httpError struct {
-	Code    int    `json:"code"`
-	Message string `json:"message"`
-}
-
-func (e httpError) write(w http.ResponseWriter) {
-	w.WriteHeader(e.Code)
-	_ = json.NewEncoder(w).Encode(e)
-}
-
-func (e httpError) Error() string {
-	return e.Message
-}
-
-var (
-	errUnknownBlock = &httpError{
-		Code:    http.StatusNotFound,
-		Message: "Block not found",
-	}
-	errBlobNotInES = &httpError{
-		Code:    http.StatusNotFound,
-		Message: "Blob not found in EthStorage",
-	}
-	errServerError = &httpError{
-		Code:    http.StatusInternalServerError,
-		Message: "Internal server error",
-	}
-)
-
-func newBlockIdError(input string) *httpError {
-	return &httpError{
-		Code:    http.StatusBadRequest,
-		Message: fmt.Sprintf("invalid block id: %s", input),
-	}
-}
-
-func newIndicesError(input string) *httpError {
-	return &httpError{
-		Code:    http.StatusBadRequest,
-		Message: fmt.Sprintf("invalid index input: %s", input),
-	}
-}
-
-func newOutOfRangeError(input uint64, blobCount int) *httpError {
-	return &httpError{
-		Code:    http.StatusBadRequest,
-		Message: fmt.Sprintf("invalid index: %d - block contains %d blobs", input, blobCount),
-	}
 }
